[RENAME] #56 - clubnotice로 이동
- 파일 폴더 변경으로 인한 package / import 변경
