+++ src/main/java/com/core/linkup/club/clubnotice/entity/ClubComment.java
@@ -1,4 +1,4 @@
-package com.core.linkup.club.club.entity;
+package com.core.linkup.club.clubnotice.entity;
 
 import com.core.linkup.common.entity.BaseEntity;
 import jakarta.persistence.Entity;
