+++ src/main/java/com/core/linkup/club/clubnotice/repository/ClubCommentRepository.java
@@ -1,6 +1,6 @@
 package com.core.linkup.club.clubnotice.repository;
 
-import com.core.linkup.club.club.entity.ClubComment;
+import com.core.linkup.club.clubnotice.entity.ClubComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
